simplify sync_committee_aggregator_proof_type tests
